Add workspace data source.
This commit adds the ability to lookup a workspace by name, and convert
all tests to use this instead of being provided a workspace ID.

The workspace ID was being looked up outside of the integration test
function, which meant running `go test` would fail if not provided valid
credentials, even when not running integration tests.

This commit also enables integration testing on jenkins run.

Change-Id: If9e7308a71cc2bdfdc7c67b6a29770274bd16bdc