Refactor all the things.
This commit is a significant refactor to address long standing issues in
how to iterate on sets of datasets in place. Unfortunately it is one
large commit due to the difficulty of breaking things up into piecewise
changes that would both satisfy jenkins and result in a usable terraform
provider which we would want to support.

The features this change supports:

- use `saveDataset` endpoint, get rid of `observe_transform` type
- recompute changes throughout the DAG
- new resource `observe_fk` for declaring foreign keys

There are three top level directories with changes:

- `vendor/` can be ignored. There is a lot of dependency bloat due to
the transitional nature of the terraform-plugin-sdk/v2 package, which
support both backward compatible and experimental changes. I expect this
to get smaller as the SDK becomes stable.

- `client/` contains code to interface with our API. `client` itself
provides a somewhat streamlined API, while code to interface with our
actual API is contained within `client/internal/api` Data sources
working.

- `observe/` contains the terraform schemas and validation.

The best place to review are the tests under `observe`, since they
contain actual working examples of terraform manifests. The terraform
model differs a bit from the object model exposed by our API:

- we compute an "oid" for every object - the reasoning is documented in
https://www.notion.so/observeinc/Terraform-OID-f6a0e00e49e345b2a9122eb006b4bae2

- a set of `inputs` is defined at the dataset rather than the stage
level. This makes it easier to visually inspect the data dependencies of
a dataset, as well as simplifying the code for recomputing changes in
the case of upstream changes.

- a stage has a pipeline, an `input` (the default input from the set of
`inputs`), and an `alias`, which maps to the input name rather than the
stageID.

Change-Id: I3cf8b883ba74654e7f930d32e98e6e7136ec4efe