Fix nil pointer dereference if dataset missing.
Our API does not return an error if a dataset does not exist. Instead,
it returns null, i.e:

{"data":{"dataset":null}}

We did not handle the case where we attempt to retrieve a dataset by ID
and it does not exist. This can happen if a dataset is deleted outside
of terraform, or local state does not match remote state (e.g.
accidentally running terraform against a different customer).

This commit makes sure we return an error in any case we attempt to
unmarshal a null value.

Change-Id: Icbd9f9301d78a0b3121ea7c5076da6c5d009d842