Add debug info on lookup failure.
Terraform doesn't currently always add the context of where an error
happened, so if you run lots of things in parallel and something goes
wrong, you're often left wondering what happened.

This commit plugs in one of those holes, adding identifier information
when we attempt to retrieve workspaces or datasets that do not exist.

Change-Id: I002e8b3e7c0c35b73f9e551468c29e5012bdb911