provider: add sweeper test
During tests, terraform may fail to clean up resources it creates (e.g.
datasets, foreign keys) if the underlying test fails in an unexpected
manner.

The terraform SDK provides the concept of a sweeper - a test which is
run specifically for the purpose of cleaning up any lingering resources.
This commit adds a sweeper test for datasets. Since our tests always use
a consistent prefix (tf-<id>) for dataset names, we can match on this
prefix to determine whether to delete the dataset.

To run a sweep:

`TEST=./observe SWEEP=all make sweep`

Change-Id: I572ec706405c59ed0487748d09f4e6708a378e67