client: add support for experimental flags.
We need a way of toggling client behavior in order to test new features,
or disable old workarounds. Rather than introduce a new provider
attribute for each potential toggle, this commit introduces a flags
string which can be passed in through the `OBSERVE_FLAGS` env var.

The format is just a comma separated list of strings, which may be
negated when prefixed with a !. For example:

`OBSERVE_FLAGS=!feature1,feature2`

would explicitly disable `feature1` and enable `feature2`. We won't
strictly check whether these identifiers mean anything: this is just a
quick and dirty way of toggling internal behavior on the fly for
experimenting. It is not intended to be customer facing, and can
therefore be relatively opaque to users.

Change-Id: I84a48b6604da9950caf34be2812a56841a6bb1fe