provider: bugfix for read missing fk
We had a potential nil reference panic when trying to look up a foreign
key that does not exist. This would happen because we would register the
error diagnostic, but not return from the function immediately after.

This commit fixes that, but also replaces most of the error handling
with diag.Errorf() function, which was recently added as a helper to the
terraform SDK.

Change-Id: I8c4ec514f60d653193474a1eef603cad8f664fdd