observe_link: ignore dataset version
Currently terraform takes into account the dataset version when deciding
whether to republish a link. However, since the version now increments
on system updates, we get the following sequence of rematerialization
hell:

1) we publish a link L from A -> B
2) B gets a new version, as do all dependent datasets
3) on next terraform iteration, L is republished, go back to 1)

Step 3 should be a noop, but in practice it leads to rematerialization
of all affected datasets. This commit avoids this by ignoring dataset
version.

 This has two side effects:

- we can no longer verify if L is valid if either A or B change.
- we do not avoid a first rematerialization after links are first
  applied. Terraform only converges after second run.

The long term fix for this is OB-2560.

Change-Id: I26afd45e8b460bda1b571231ed40cf45117f7d81