Modify upload directory structure.
This is a necessary step for bringing our provider in line with the
recent changes in Terraform 0.13. We now keep the same directory
structure as releases.hashicorp.com, that is:

${PROVIDER_NAME}/${VERSION}/${PROVIDER_NAME}_${VERSION}_${OS}_${ARCH}.zip

When pushing to S3, we no longer provider a "latest" helper. This was                                                                                                             intended as a convenience method for customers, but in practice we're
going to skip it all together by implementing a third-party Terraform                                                                                                             provider registry, so things will "just work" in terraform 0.13 with the
following snippet:

```
terraform {
  required_providers {
    observe = {
      source  = "terraform.observeinc.com/observeinc/observe"
      version = "~> 0.3.0"
    }
  }
}
```

Change-Id: Id50b0de253ce76b633ea9977d5e095b2d135f82d