jenkins: add ability to sign provider plugins
As part of process to be a host a third party provider registry, we need
to sign our plugins. This involves taking a shas of the outputs, and
signing the resulting file.

I've uploaded a private key that we can use for now, still need to give
jenkins access to it.

Change-Id: I4794df522016d902300184c93171bad219936b31