Split out API by type.
Previously we had a single internal API package, which corresponded to
the v1/meta GraphQL endpoint. Over time, we've had to interact with
RESTful endpoints too: first with /v1/login, and soon we will need to
support the collector API. This ends up being a bit of a mess, not only
because of graphQL vs REST, but also because of the different domain
names.

This commit formally breaks the API into subpackages under the client,
so we can more clearly separate how each one behaves:

    - internal/meta: customer GraphQL endpoint mounted under /v1/meta,
    - internal/customer: customer REST client mounted under /

In an upcoming change, I will add:

    - internal/collect: collector client

Change-Id: I7d41fa8b9cacf5c0945ea8de33f8c0e810acfc1a