Implement collect API, streamline client configs
This commit adds the collect API. As a side-effect, I decided to
cut our losses and drop our use of functional options in configuring the
client.

The client config was getting cumbersome - in part because we were using
decorators around the RoundTripper interface. These would work great if
we didn't occassionally need to access the underlying http.Transport
(for configuring TLS, adjusting proxy). Once you have arbitrarily
wrap RoundTrippers, you lose some control - for example, we want to
log requests, but we don't want to leak sensitive information from
the /v1/login endpoint in particular.

The resulting code is a bit cleaner - client.Client is responsible for
configuring an appropriate HTTP client which handles logging and auth.
This *http.Client is then passed to the clients for each individual API.

Change-Id: Icead4ced760fe7b073b24b42ed6d9978d8b647ff