Add observe_http_post resource
As part of attempt to use Terraform for some more formalized e2e
testing, we need to be able to submit data, and then query it. This
commit handles data submission.

We add a new "observe_http_post" terraform resource which takes input
string data, and sends it to our collector. At it's simplest:

```
resource "observe_http_post" "test" {
  path = "/terraform/test"
  data = jsonencode({ "hello" = "world" })
}
```

Obviously this doesn't map directly to a resource in a Terraform sense:
there's no practical state we can query in a meaningful timeframe, and
even if there were, there's no way to "update" it. We're abusing the
model slightly to get something useful.

When this resource is created, we generate a unique ID, which gets
submitted as a tag alongside our observations. By omission, we use
"tf_id". As a result, we can configure a dataset to filter out submitted
observations, e.g:

```
resource "observe_http_post" "test" {
  path = "/terraform/test"
  data = jsonencode({ "hello" = "world" })
}

...

resource "observe_dataset" "my_observations" {
  workspace = data.observe_workspace.kubernetes.oid
  name      = "Submitted data"

  inputs = {
    "observation" = data.observe_dataset.observation.oid
  }

  stage {
    pipeline = <<-EOF
      filter EXTRA.tf_id = "${observe_observation.test.id}"
    EOF
  }
}
```

Finally, we can configure the observation resource to expire after a
given "refresh" duration. If this period has elapsed, on a subsequent
Terraform run we will resubmit the observation data with an entirely new
ID. This gives us a quick and simple way of building "static resources",
for which we currently use cronjobs to submit data periodically (think
HTTP headers in cars-app).

Change-Id: I6fa1cff4ea6d7fd18ad6ff848934c91a137da68e