provider: add observe_channel resource
This commit adds observe_channel. This is relatively straightforward,
the only novelty here is that we're passing list of objects (channel
actions), rather than creating each relationship via a separate type (as
we do with observe_link). The risk of this approach is that any user
added actions to an existing channel managed by terraform will get wiped
out on next terraform apply.

Change-Id: I607d98e38939517b043c07cb1f3033018027cd12