provider: add polling to `observe_query`
In some cases we submit data, and we need to be able for it to be
available in the Observation table. To allow for this, we add a polling
capability to our data source: given an interval and a timeout, we will
retry our query until we get rows back.

One caveat here is if the user did not specify an end time, we must
recalculate "now" on every iteration of the query.

Our data source test now submits an observation, and waits for it to be
available! In this case I use the default timeout and interval options.

Change-Id: I49b3a5cb866a0de600b3c2270d5b040c14648105