provider: add assert option to observe_query data source
The final requirement for testing is being able to compare a query
result to some reference point. This is thorny since terraform doesn't
have great primitives for comparing data structures, and because we
can't coldrop "valid to" and "valid from" timestamps, which are
inherently tricky to compare.

This commit introduces an optional `assert` option, which allows users
to compare a query result against a golden file. We ignore timestamp
columns in our comparisons for now. I've used go-cmp because I expect
diff behavior might be configurable in the future, and I want to retain
some flexibility in how this comparison is handled. Hopefully we don't
need much in practice, and we mostly rely on OPAL to do the heavy
lifting.

Change-Id: I8d5bdefdeba202dd2b4f7697faf328179278aeca