provider: remove `refresh` attribute from observe_http_post resource
The `observe_http_post` resource included a `refresh` attribute which
would allow the resource to be periodically recomputed. This is useful
in order to ensure that data is reuploaded if it has not been submitted
after a given period. For example, the following snippet would resubmit
a JSON if over an hour had elapsed since the previous apply:

```
resource "observe_http_post" "test" {
  data    = jsonencode({"hello" = "world"})
  refresh = "1h"
}

This functionality can be completely replicated using the time module.
This is a bit more verbose, but more composable and flexible:

```
resource "time_rotating" "timestamp" {
  rotation_minutes = 60
}

resource "observe_http_post" "test" {
  data    = jsonencode({"hello" = "world"})
  tags = {
    "ref_timestamp" = time_rotating.timestamp
  }
}
```

`time_rotating.timestamp` will be recreated once the rotation period has
elapsed, which will trigger a new HTTP POST.

Change-Id: I394f8cda945ab589a3987ab53641ed2b105a3c8c