Add preliminary monitor support
For now we only support CRUD-ing a monitor, and two rule types (Change
and Count). I'll add more rule types incrementally in order to not make
this more bloated.

Review suggestion is looking at the test in
observe/resource_monitor_test.go, and then the schema in
observe/resource_monitor.go. Almost everything else is boilerplate
conversion of stuff, of which there is a looot of.

Most of the code is janky, brittle conversions because once again we're
dealing with unions, and there's a lot of boilerplate derived from
supporting string enums. The main sources of bloat:

- input and output structs are always independent entities
- terraform expects schema keys to be snake case, our API uses camel case
- extend that expectation to enums, for stylistic consistency
- handling custom types such as NumberScalar
- trying to be helpful by providing "value" to encode compare functions
  that only require one value, vs having users write an array every time
  in "values"

Change-Id: I7011a2bbfcecaf05246ce62674f98fe4798c2712