client: add source identifier
This commit adds a source identifier to datasets managed by terraform.

This identifer is opaque to Terraform users: we do not surface is as
part of the configuration, and the dataset source field will get
blindly overwriten. The source category is always "terraform". The
source comment can be configured as part of the provider, but will by
default match the user_email, if any exists.

Change-Id: I392c89b32e37ab9494f49be3f16c4f560f20f16e