dataset: use lastSaved as version
In the API, dataset `version` tracks all changes to the dataset object,
including those that are carried out as a result of introducing links.
This throws off Terraform, that needs a version identifier to determine
whether it has to recompute downstream objects. Instead of using system
version, we modify the client code to rely on `lastSaved` instead, which
tracks the timestamp of the last time a dataset modification was
the result of a saveDataset call.

Change-Id: I95cfb067f00e352712f1e43c776fd504f32844f4