dataset: suppress pipeline diff for trailing whitespace
Editing a dataset in the UI can lead to pipeline being stripped of a
newline, or users inadvertently adding a bunch of additional whitespace.
This can lead to a confusing diff presentation from Terraform, where a
change is detect, but no actual diff is highlighted.

This commit ignores any diffs introduced by trailing whitespace. Since
the OPAL is functionally the same, we can ignore trailing whitespace
safely.

Change-Id: I0eb4fe5dd2153f359366ff097ec7b1743501bca4