observe_dataset: ignore stage foreign keys
A foreign key may target a stage, rather than a dataset. We don't really
handle this, because such a relationship cannot currently be described
in a terraform manifest. However, if a user edits an dataset managed by
terraform, it is possible such a key gets introduced and causes the
provider to crash on a nil pointer.

Change-Id: I83efb66c15842d5d3da0a103d8c7c43728d8381f