Make TestAccObserveSourceDatasetResource less flaky
This test will fail if a dataset with the same source table already
existed beforehand, which could happen if the test command was
cancelled early. If the source table existed beforehand, the
test would fail and indicate that the platform_table_uniqueness_constraint
was voilated.

This change makes it so that the source table names now depend on
the name of the dataset, so the uniqueness constraint only matters
if the random prefixes are the same (which is very unlikely)

Change-Id: Ieada935254da0327336c8a3c5b2ee59834fab627