fix: handle resource diff prior to resolution
We need to take into account that a value may contain a reference to
something that is not yet known. This is common in boards, where the
JSON payload will embed an ID which hasn't yet been created. In such
cases, we need to make sure we correctly correct a diff between objects,
rather than silently ignoring unmarshaling errors and assuming value
equality.

Change-Id: I9821e4b74ed571f178fd8215732b4f53118d1cb8