fix: ignore group_by_columns if group_by_group is used
When setting group_by_group, the backend reuses the columns value to
populate group_by_columns. This causes a perpetual diff between desired
config and actual state. Since the group_by_columns field is on the way
out, just work around this limitation by not using it if
group_by_group is set.

Change-Id: I8bca7dd19c9d40033530b29f08a2402b532aa73b