fix: elide path_cost if absent
This commit restores behavior prior to 0.10.0, whereby we'd elide the
`path_cost` from state if not provided.

I added tests to verify this behavior, but I'm not sure whether to stick
with it or not. The implementation is a bit weird, and I suspect there
are edge cases where it might surprise us? On the other hand, going with
the more straight forward solution causes every single dataset to have a
change, which is panic inducing for users.

Change-Id: I8bec91c4d46dde101d917dd0e50c6d68e2dddc14