fix: separate folder from workspace
Previously we'd allow a workspace to be used in the folder field. This
would cause a perpetual diff, since the API would create a folder for
us. The actual solution is to allow `folder` to be a computed field if
not set.

Note that this change is technically backward incompatible, since
workspace will no longer be allowed for the folder field. Not sure we
use this in practice however.

Change-Id: I82286c70125487ecd20a4abf0f0b29532d6b3bb9