fix: remove datasets from workspace lookup
We previously were listing all datasets in a workspace whenever we did a
workspace lookup. This turns out to be a pretty bad idea - the number of
datasets can grow unbounded, so things can get slow and we will
eventually need pagination. Remove this feature before anyone relies on
it.

Change-Id: I63b303800a6878d795fab87ccb01f34fdcd29c2d