fix: send null request method rather than empty string
When setting the `method` field for the HTTP poller, our provider was
sending an empty string rather than a null. This became problematic
once we switched to doing regex based matching on the match stanza. As a
result,

rule {
   match {
      url = "https://example.com"
   }
}

would never match, since the crawler client would be applying a
regex_match(request.Method, /^$/).

This commit is largely boilerplate-y in nature. Concerningly, it's
really hard to get a test that captures the underlying change, since
terraform-plugin-sdk makes it horrifically hard to distinguish between
empty string and null. For now, just treat empty string as a null, and
make sure we accept any regular expression when performing a match.

Change-Id: I28f2398bf2313346fdfacad6d5818ea8521ea9dd