preferred_path: allow "error" status on read
When a preferred path is read, it may be in an error state. This can
occur due to changes to related objects, such as links. While this
should trigger an error on create/update, it should not prevent a read
(refresh) of the resource.

For now, this surfaces the error via a warning diagnostic. This should
generally be visually presented to the user, but not cause the Terraform
run to fail. Optionally, we might also expose the error to users as a
resource attribute in a future change. This would allow users to target
the error with a postcondition:

https://developer.hashicorp.com/terraform/language/expressions/custom-conditions

OB-13810

Change-Id: I3fe30dc01ec9dcecdada981750a243297d0bc12f