test: use -json to prevent output buffering
When testing multiple packages, "go test" buffers all the output until the tests finish, even with verbose (-v) mode. This is because parallel tests could otherwise interleave their output and be difficult to read. Across packages, tests could have the same name.

In JSON mode (-json), there's an explicit "Package" field in the output, so "go test" can safely print output as it comes in.

This ensures that if the tests are running for a long time due to server errors, the user can see the output as it comes in. Otherwise, Jenkins may terminate the job at the configured 40m deadline, before any meaningful output is printed.

https://github.com/golang/go/issues/46959

Change-Id: Ia6e0bf92e927b14f56e601839b27141bb6dd1298