make: build static binary without cgo
When running in Alpine Linux, such as in the hashicorp/terraform Docker image, the provider currently fails to start:

╷
│ Error: Failed to load plugin schemas
│
│ Error while loading schemas for plugin components: Failed to obtain provider schema: Could not load the schema for provider terraform.observeinc.com/observeinc/observe:
│ failed to instantiate provider "terraform.observeinc.com/observeinc/observe" to obtain schema: Unrecognized remote plugin message:
│
│ This usually means that the plugin is either invalid or simply
│ needs to be recompiled to support the latest protocol...

This failure is the result of a dynamically linked binary. Executing the provider directly should print:

This binary is a plugin. These are not meant to be executed directly.
Please execute the program that consumes these plugins, which will
load any plugins automatically

Instead it currently prints:

qemu-x86_64: Could not open '/lib64/ld-linux-x86-64.so.2': No such file or directory

Disabling cgo is standard for building providers:

https://github.com/hashicorp/terraform-provider-scaffolding/blob/8450b2b3a4ae1dc5b124bdfa1c065a6077408eba/.goreleaser.yml#L12

Change-Id: I3854651a840e3c13e24a0ec1a7fa8051fb57c056