monitor: implement `comment` field for data/import
* Adds `comment` field to `data.observe_monitor` (with test coverage)
* Handles `comment` when `Read` method of monitor resource is called (shared with data source)
  * Updates GraphQL fragment to include `comment`, re-generate client with `go generate`
  * Because the field is nullable at the API level, it's handled slightly differently than other fields
  * Even if unshared, this would be necessary to provide drift detection and support `terraform import`
  * If this were included when `comment` was introduced to the resource the data source test should've failed

This is a follow up to https://gerrit.observeinc.com/c/terraform-provider-observe/+/27657, where I was the reviewer and didn't catch this.

As part of my research on Observe's "Export as Terraform" feature, I found that data sources and resources for a given object must be equal. Any fields not present in the data source will not be "exported" (generated).

https://www.notion.so/observeinc/Export-as-Terraform-v2-1a6308286bee4973adf5bc8786deadd8#16c69ae0391f4bb786a0ee48489e4d2f

That research was what prompted me to think about how `comment` didn't adhere. And now that I'm doing this one, there's surely others.

https://gerrit.observeinc.com/c/terraform-provider-observe/+/25752 was another starter task and has the same bugs. I can fix that next but we also need better testing/guardrails.

The recently added requirement to generate docs as part of each PR should help make this a bit easier to spot:

https://gerrit.observeinc.com/c/terraform-provider-observe/+/28416

Change-Id: I677a309f1a6e9781985b69caac2a542467df9005