dashboard: fix corrupted parameters defaultValue when reading from API
The API sends the following as the `defaultValue` for a null parameter of type bool:

  {"bool":null}

The generated GraphQL client does not handle this. When a dashboard is round-tripped, the data becomes the following:

  {
    "bool": null,
    "float64": null,
    "int64": null,
    ...
  }

The API rejects this value:

  input: dashboard.dash.parameters[0].defaultValue ValueInstance JSON object cannot be unpacked: map[string]interface {} (map[])

The generator (genqlient) cannot handle this type because it has an implicit behavioral requirement that cannot be directly represented in Go.
The fields in the generated type backing these objects are all pointers, but those can only be defined or nil. Go cannot represent unset vs explicit null for a struct field.

As a workaround, because we're not actually interacting with the content of Parameter objects let alone the fields, we can use a JsonObject type which stores the data as a string and is guaranteed to idempotently round-trip.

This is a shortcut to avoid defining an entire struct type in Go that can handle this API type. If in the future we had `parameter {}` blocks in dashboards and a similar structure for accepting defaults of different types, we'd need to define that type.

https://observe.atlassian.net/browse/OB-14005

Change-Id: I0ee07baa7c3af38e0d262b4b6675112789cf016a