dashboard: implement custom Go type for handling parameter defaultValue
Adds a custom Go type for handling the defaultValue field of a
parameter. This type mimics the behavior on the server can handle null
values. It provides greater backward compatibility than just sending the
raw JSON, by stripping extra nulls:

A user can specify:

  defaultValue = {
    bool = null
    string = "foo"
  }

Previously, an API bug treated this as a null bool. This has been fixed:

https://gerrit.observeinc.com/c/observe/+/30010

This fix has not been released to production yet.

This applies the same logic on the client side, stripping nulls from the
configuration before sending JSON to the server.

Users will no longer see perpetual diffs for extraneous nulls, because they
are stripped when unmarhsaling and re-marshaling the user's config.

https://observe.atlassian.net/browse/OB-16140

Change-Id: Ieba8bf2c50d0128b4d111ae99ff0b80128d68161