monitor: add support for setting reminders
Monitors have a `notification_spec` that can enable periodic reminders during the triggered period.
This adds support for setting that via Terraform.

The API for notifications is a bit awkward. When notifyOnReminder=true,
no default is set for `reminderFrequency`. In effect, you must pass both fields to achieve the intended result.

Expected behavior would be a default from the API or an error.

The UI sets a default of 10m, but this is buried deep within its
implementation of a `Duration` type and not formalized anywhere.

SDK v2 does not support requiring a field to be set when another field
has a certain value (on when it is set at all).

A frequency should not be required:

  notification_spec {
    notify_on_reminder = false
  }

A frequency should be required:

  notification_spec {
    notify_on_reminder = true
  }

Instead, I've implemented `notify_on_reminder` as a computed attribute.
The user sets `reminder_frequency` and the provider handles enabling
reminders if the frequency is non-zero.

https://observe.atlassian.net/browse/OB-16177

Change-Id: Id9f18a3e64fcd9f60d38602a38a9b374059561ed