dashboard: fix defaultValue.primaryKeyValue[].name
The GraphQL type of primaryKeyValue entries is "ValueKeyValue"

But its fields are not "key" and "value" but rather **"name"** and "value"

We had no test coverage against this parameter default type before, so it slipped through.
I've added an acceptance test covering this case here.

In the absence of this fix, the new test fails with the reported error:

  terraform-provider-observe/observe/resource_dashboard_test.go:1279: Step 2/2 error: Error running apply: exit status 1

  Error: failed to update dashboard [id=41252249]

    with observe_dashboard.first,
    on terraform_plugin_test.tf line 15, in resource "observe_dashboard" "first":
    15: 				resource "observe_dashboard" "first" {

  returned error 422 Unprocessable Entity: {"errors":[{"message":"must be defined","path":["variable","dashboardInput","parameters",2,"defaultValue","link","primaryKeyValue",0,"name"],"extensions":{"code":"GRAPHQL_VALIDATION_FAILED"}}],"data":null}

https://observe.atlassian.net/browse/OB-16421

Change-Id: Ife7ddbadb26d6515d7a784ddc7d6b9c153f8900a