dashboard: marshal float64 as float64 instead of string
With Int64Scalar, the API marshals the value into a string, so it makes
sense for the provider to match.

With NumberScalar, the provider is behaving differently than the API:

  func MarshalNumberScalar(n NumberScalar) graphql.Marshaler {
    return graphql.WriterFunc(func(w io.Writer) {
      if n == "" {
        //	the representation of the zero-initialized number is 0
        w.Write([]byte{'0'})
      } else {
        w.Write([]byte(n.String()))
      }
    })
  }

We should not be wrapping every float64 in a string, even though the API
will handle it. Arguably, we shouldn't be doing that for int64 either in
a language with support for 64-bit integers, but that's a potential
future change, for now we can match the API.

Change-Id: I13ab754a3d276b196df1b9a13f8055112a70df1f