fix: restore marshaling of floats as strings
While the API's UnmarshalNumberScalar method is apparently written to
handle actual numbers (int64, float64), in reality the JSON decoder is
configured to decode these as json.Number. This requires no handling and
can be returned as-is, since it's already the target return type. But
it's not handled, and thus falls through to the default case and errors.

Until this bug is fixed and valid JSON numbers are accepted as input, we
need to continue sending them as strings.

Otherwise, the API returns an error:

  input: monitor.monitor.rule.thresholdRule.compareValues[0] UnmarshalNumberScalar() for json.Number not implemented in metatypes

Change-Id: I99298bec47394baafc910b24215301b91f4f42c1