workspace: avoid fetching datasets when getting workspace
data.observe_workspace is heavily used to look up the default workspace by name.
Each time, it fetches all datasets in the workspace too.

This was at one point part of the schema, but was removed in:

https://gerrit.observeinc.com/c/terraform-provider-observe/+/25248

Now, the only dependency was the dataset sweeper, which is updated here to make a second query.

https://observe.atlassian.net/browse/OB-15493

Change-Id: I8dca1e3d9c642b9612ebaa1560faad06ebac8e3d