fix: adjust SaveMode in dataset
When migrating to gengql we inadvertently changed the "save mode" of a
dataset from `SaveModeUpdateDatasetAndDependenciesIgnoringAllErrors` to
`SaveModeUpdateDataset`.

Surprisingly this didn't have much impact since the API appears to not
be very strict. We are now adding more backend validation around this,
so it is good form to convey that we don't care that we break downstream
datasets in practice.

Change-Id: I16f18ef4dc771073165f7d0aa15d67e3121541ef