ci: pass in pattern to sweeper
Our current sweeper takes a "sweep" argument and does nothing with it.

Since we consistently check for the name of the object being deleted, it
makes sense to pass in the regular expression rather than hard coding it
in the sweeper code. In this manner, we could re-use the sweeper to
delete objects from our other CI workspaces (content) by adjusting the
desired pattern.

The one caveat is that passing regular expressions through Makefile
is really tricky. We would probably want to pass the pattern in some
other way (directly via go test).

Change-Id: Ic8e8a82ab6ccd69fcab0cb07190082d62eb54906