dataset: fix overly specific error matcher in test
The existing error matcher attempts to match the entire error text,
making it brittle to API changes and whitespace-dependent. This test is
currently failing due to an upstream change.

The new error:

  input: dataset errors in stage "stage-0": 1,14: [EXTRA] ...

The expected error:

  ...: 1,14: [] ...

In effect, the old error message had a bug, which is now fixed, and
the provider tests depended on that bug.

Instead, this changes the matcher to the minimum necessary regex to
reasonably match this type of error (missing field, EXTRA) without the
brittleness of asserting its exact content, which the provider doesn't
directly control.

Change-Id: Ifff8062989ee05c2e4cbad7b7867b4a4d6f1868b