release: remove `v` prefix from version in Boring Registry (S3)
Boring Registry expects bare semver versions, without the `v` prefix, in its `version=${version} key segment.

https://github.com/observeinc/boring-registry#provider-registry-protocol

This workflow current includes git tag, which includes the prefix:

https://github.com/observeinc/terraform-provider-observe/actions/runs/5612625311#summary-15206981362

The Terraform Registry requires tags with the `v` prefix and generally prefixing semver tags is advisable. So we need to switch, but just parse the tag to get the version component, sans prefix.

The action we're already using to validate the semver tag parses it to do so and exposes the different properties as outputs. So I've exposed that from the `goreleaser` job so that the `release` job can use it.

In theory tag validation could be its own job that runs before `goreleaser` and then the `release` job could just use the output from that. But ultimately that just takes longer as it incurs overhead for job startup, even if it doesn't have to check out the code. So I've deferred factoring that out until there's a more complex graph, with something running in parallel to `goreleaser`.

Change-Id: I331e68be4eeb475c6da97ff96e5ec59b94b2c5e4