fix: ensure lists of objects are exported as blocks for monitor v2
Currently, when exporting a TypeList with Elem: schema.Resource, it gets exported as a list of objects (using square brackets) instead of blocks, which is invalid.

Turns out the `terraform show` command used by our export implemention will have this behavior when `Computed: true` is used without either `Optional: true` or `Required: true`. This is likely because the block format only makes sense when it's possible to use the field as an input. For purely computed fields, we would only reference the data and the list format is equivalent in that case.

Added `Optional:true` to all relevant fields, though it is not actually valid to provide these fields as input when querying and will be ignored.