feat: Fail terraform apply if edit-forward cannot be applied
Currently if a dataset cannot be edited-forward, it will be rematerialized (best-effort behavior). This is an unpleasant surprise for users because there is no mechanism use dry-run in Terraform (unlike the WebUI). This change fails those changes, and the user should either make their change compatible with edit-forward or remove the rematerialization_mode flag