fix: re-create links if not found during update
Normally, when running terraform apply, it'll refresh the state first at which point if a link no longer exists, terraform will plan to re-create it automatically. However, in the case where the link does exist when refreshing the state, but then is deleted during the apply (e.g. due to the source dataset being updated), the update to the link then fails. We now will re-create the link if the update fails due to link not found.