fix: re-create dataset query filters when the dataset id changes
See comments on the ForceNew field for details.

There is still one rare case that causes issues, namely when you create a new dataset (i.e. the dataset id, not just the version, is unknown during the plan stage) AND update a dataset filter to reference that new dataset. In this case, the CustomizeDiff will not apply since the new dataset id value is "(known after apply)" so terraform will plan an Update, but then during the apply stage, it'll realize that this should have been a Delete + Create and error with "Provider produced inconsistent final plan". But this case should be quite rare and solvable by simply re-running the terraform apply (it'll pick up where it left off and this time the new dataset id is known, so the CustomizeDiff and ForceNew will apply as desired).