feat: add server-side validation of datasets during plan stage
Abuses CustomizeDiff to make a dry-run SaveDataset call during the plan stage when certain dataset fields are modified. If there's an error, or if there would be rematerialized datasets and rematerialization_mode=must_skip_rematerialization, fails the plan stage.

Limitation: If any fields are unknown during the plan stage (e.g. it's referencing a new dataset that's going to be created during this terraform run), we cannot properly call SaveDataset. So we skip server-side validation in such cases.