fix: add missing diff supress for enum fields
Currently although we expect snake-case enums, the validation is case insensitive meaning something like "Email" is valid. But we set it to the standardized lower-case "email" in the state, resulting in a perpetual diff. Turns out there's already a `diffSuppressEnums` for this issue, but it just wasn't being used in many places.